Coupon dialog: move NITWKWebViewContainer to its own file
